Gene Set Adaptation
Function that uses a selected cohort and customized signatures for igraph analysis. This function returns a dataframe with the information of adapted signatures. The adaptation is based on CCIM and NCEC value. Moreover, from selected path, gsadapt function save two igraph PDFs with the information of spatial distribution of each genes tested in the signature.